// apps/web/src/api/models/PoolDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PoolDto = {
    /**
     * ID of the pool
     */
    id: string;
    /**
     * Name of the pool
     */
    name: string;
    /**
     * Address of the pool
     */
    address: string;
    /**
     * Details about the pool
     */
    details: string;
    /**
     * Image URL of the pool
     */
    imageUrl: string;
};